Move the implementations of the various plugin convention objects into internal packages and leave behind abstract classes in the public API.
