+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/ApplicationPlugin.java
@@ -27,13 +27,12 @@ import org.gradle.api.distribution.DistributionContainer;
 import org.gradle.api.distribution.plugins.DistributionPlugin;
 import org.gradle.api.file.CopySpec;
 import org.gradle.api.internal.IConventionAware;
+import org.gradle.api.plugins.internal.DefaultApplicationPluginConvention;
 import org.gradle.api.plugins.internal.DefaultJavaApplication;
 import org.gradle.api.tasks.JavaExec;
 import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.Sync;
 import org.gradle.api.tasks.application.CreateStartScripts;
-import org.gradle.internal.Factory;
-import org.gradle.util.DeprecationLogger;
 
 import java.io.File;
 import java.util.concurrent.Callable;
@@ -106,12 +105,7 @@ public class ApplicationPlugin implements Plugin<Project> {
     }
 
     private void addExtensions() {
-        pluginConvention = DeprecationLogger.whileDisabled(new Factory<ApplicationPluginConvention>() {
-            @Override
-            public ApplicationPluginConvention create() {
-                return new ApplicationPluginConvention(project);
-            }
-        });
+        pluginConvention = new DefaultApplicationPluginConvention(project);
         pluginConvention.setApplicationName(project.getName());
         project.getConvention().getPlugins().put("application", pluginConvention);
         project.getExtensions().create(JavaApplication.class, "application", DefaultJavaApplication.class, pluginConvention);
