+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/WarPluginConvention.java
@@ -17,48 +17,24 @@
 package org.gradle.api.plugins;
 
 import org.gradle.api.Project;
-import org.gradle.util.DeprecationLogger;
 
 import java.io.File;
 
 /**
  * <p>A {@link Convention} used for the WarPlugin.</p>
  */
-public class WarPluginConvention {
-    private String webAppDirName;
-    private final Project project;
-
-    /**
-     * Creates a {@link WarPluginConvention}.
-     *
-     * @deprecated Creating instances of this class is deprecated. These should be created by the WAR plugin only.
-     */
-    @Deprecated
-    public WarPluginConvention(Project project) {
-        DeprecationLogger.nagUserOfDeprecated("Creating instances of WarPluginConvention");
-        this.project = project;
-        webAppDirName = "src/main/webapp";
-    }
-
+public abstract class WarPluginConvention {
     /**
      * Returns the web application directory.
      */
-    public File getWebAppDir() {
-        return project.file(webAppDirName);
-    }
+    public abstract File getWebAppDir();
 
     /**
      * The name of the web application directory, relative to the project directory.
      */
-    public String getWebAppDirName() {
-        return webAppDirName;
-    }
+    public abstract String getWebAppDirName();
 
-    public void setWebAppDirName(String webAppDirName) {
-        this.webAppDirName = webAppDirName;
-    }
+    public abstract void setWebAppDirName(String webAppDirName);
 
-    public final Project getProject() {
-        return project;
-    }
+    public abstract Project getProject();
 }
